Read the file to be uploaded
Begin the upload script by testing to make sure we can get
the filename and some data back from fs.readfile.

Use an empty string as a default argument for the file path so
that if we don't pass an argument in running the script, we get
a sensible error message (file not found).